Added scripts for processing pollinator videos
Added preliminary script for helping to automate processing of
pollinator videos. Script accepts video files, allows for selection of
an area containing time stamp info, and matches that area to a supplied
reference file to parse the digits by template matching to the reference
file.

A jupyter notebook used to prototype the parsing functions is also
included.

Signed-off-by: Avery Uslaner <uslaner.avery@gmail.com>